+++ magick/gem.c
@@ -736,16 +736,16 @@ MagickExport PixelPacket InterpolateColor(const Image *image,
     return(image->background_color);
   p=image->background_color;
   if ((x >= 0.0) && (y >= 0.0))
-    p=AcquireOnePixel(image,(int) x,(int) y,exception);
+    p=AcquireOnePixel(image,(long) x,(long) y,exception);
   q=image->background_color;
   if (((x+1.0) < image->columns) && (y >= 0.0))
-    q=AcquireOnePixel(image,(int) (x+1.0),(int) y,exception);
+    q=AcquireOnePixel(image,(long) (x+1.0),(long) y,exception);
   r=image->background_color;
   if ((x >= 0.0) && ((y+1.0) < image->rows))
-    r=AcquireOnePixel(image,(int) x,(int) (y+1.0),exception);
+    r=AcquireOnePixel(image,(long) x,(long) (y+1.0),exception);
   s=image->background_color;
   if (((x+1.0) < image->columns) && ((y+1.0) < image->rows))
-    s=AcquireOnePixel(image,(int) (x+1.0),(int) (y+1.0),exception);
+    s=AcquireOnePixel(image,(long) (x+1.0),(long) (y+1.0),exception);
   x-=floor(x);
   y-=floor(y);
   alpha=1.0-x;
